simple stack example
proof of concept stack in chip8. in memory there is a stack pointer and a
pre-allocated section for the stack. a global `push` and a global `pop`
exist that store/restore the current state of v0-vd on the stack and
increment/decrememnt the stack pointer by 14 (14 register values are
saved).

`pop` returns early if the stack pointer is 0 so that it doesn't try to
decrement the stack pointer into invalid memory. currently the
equivalent does not exist for `push`.

I can only save 14 registers because in order to perform arithmetic on
the stack pointer I need at least one register that isn't vf, and
since vf is changed in this process it can't be saved either.