guard against a stack overflow
this requires testing against a predefined stack limit. in order to save
as many registers as possible with a push, `ve` is used to load the
stack pointer which means that I can't do a psuedo comparison against
the stack limit. this means that if the stack is ever increased *above*
the limit then this guard will break. if we want to support pushing an
arbitrary number of registers to the stack then we can't support pushing
v0 - vd here.